Third commit

+++ src/main/cpp/t02_dist.cpp
@@ -22,7 +22,7 @@
 using namespace std;
 
 //function distance
-int distance(double x1, double y1, double x2, double y2){
+double distance(int x1, int y1, int x2, int y2){
 	return sqrt(pow(x2 - x1, 2) + pow(y2 - y1, 2));
 }
